fix(sql-literal-typing-1): Spark's integral literal typing and promotion on the SQL door — BL-20, one Rust analyzer rule, both doors agree (#656)
* test(sql-literal-typing-1): red-first BL-20 pins over the 64-cell Spark oracle

19 failed, 44 passed on base 33c87cbf; every in-scope defect cell red,
every already-green cell pinned through to_arrow()/typeof.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* feat(sql-literal-typing-1): Spark integral literal typing before TypeCoercion

SparkIntegralLiteral narrows unsuffixed Int64 literals to Int32, types
UInt64 as Decimal128 and refuses past precision 38; installed before the
first TypeCoercion so DataFusion's own coercion yields Spark promotion.
Pins 63 green; probe 30 DIFFs down to 14 (residues only).

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* docs(sql-literal-typing-1): BL-20 FIXED, BL-20-OVF residue row

BL-20 carries the pin file, fixture and oracle path; BL-20-OVF backlogs
narrow-integral overflow with the seam named; div and unary tilde named
as SPARK-SQL-GRAMMAR-1 residues.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* test(sql-literal-typing-1): seat the fourth pre-coercion rule in the order pin

The analyzer-order contract test kept the old three-seat sequence; the
literal rule is the fourth seat. Name kept, behavior pins unchanged.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* fix(sql-literal-typing-1): facade fallout — lambda, union, overlay, TY-7

Early narrowing left two cements for TypeCoercion: re-resolve lambdas and
re-derive union schemas from narrowed inputs. Overlay -1 sees through one
compiler cast. TY-7 twins and two width pins move to Spark's recorded
answers; two fnp8 dispositions re-measured to the oracle schema.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* test(sql-literal-typing-1): cross-door bodies spell literals BIGINT, BL-20-NATIVE row

The two schema-equality tests keep testing CTAS/MERGE mechanics with
explicit widths; the native-door narrowing remainder is BACKLOG with the
DAG home question named.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* test(sql-literal-typing-1): flip incidental Int64 pins to Spark-true Int32

Literal-built fixtures now answer INT; the width halves of these pins move
with them (purposes untouched), and the cache byte budget re-measures
312 to 300.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* docs(sql-literal-typing-1): final gate lines in the ledger

All gates green: verify, both Python suites, probe scoreboard.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* test(sql-literal-typing-1): red pins for the LIT2 cells and the F.expr door

10 fail on 996094c8: LIT2-SQL-00/01/02 (Exact Int32 UDFs refuse Int64),
LIT2-SQL-03 (parenthesized INT_MIN folds to int, Spark says bigint),
LIT2-PY-00..04 plus the critic L-001 door-agreement table on F.expr.
LIT2-SQL-04/05/06/07 already green and guard the fix.

pins: sql-literal-typing-1/L-001, L-002, L-003

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* fix(sql-literal-typing-1): close L-001/L-002/L-003 on both doors

L-001: build_expr_context seats SparkIntegralLiteral through the shared
insert function, now first among the pre-coercion rules, and installs the
post-coercion rules minus the subsumed late SparkIntegerLiteral, so F.expr
types mixed-width SQL exactly like spark.sql.
L-002: DateOffset (date_add/date_sub) and Factorial shadows accept every
integer width in coerce_types and delegate kernels upstream; plan
construction verifies UDF arguments before any analyzer rule runs, so the
narrowing seat alone could never reach them. The 2-arg dateadd route now
targets the shadow.
L-003: the Negative fold is gone; only the lexer-level negative token
narrows, so -(2147483648) stays bigint while - 2147483648 stays int.
76/76 pins green, lambda/union/typeof neighbors green.

pins: sql-literal-typing-1/L-001, L-002, L-003

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* perf(sql-literal-typing-1): skip the literal walk when no literal can narrow

PERF-001: plan-level apply pre-check over Int64/UInt64/Decimal literals;
recompute_schema and resolve_lambda_variables run only after a real
transform, the latter only with a higher-order function on the node.
PERF-002: rewrite_values applies over &Expr first, clones rows only on
change. PERF-003: Union rebuilds only on a moved child schema type.
Release-native medians: 200 withColumn F.lit+schema 4454 -> 4282 ms,
F.expr+schema 4583 -> 4293 ms; (a)/(b)/(c)/UNION int-vs-str deltas stay in
the noise. 76/76 pins plus lambda/union neighbors green on the new build.

pins: sql-literal-typing-1/L-001, L-002, L-003

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* fix(sql-literal-typing-1): always recompute schemas; facade factorial uses the shadow

The PERF-001 recompute skip kept construction-time parent schemas after
bottom-up narrowing, so lambda variables bound the stale Int64 width and
24 FNP-8 HOF pins answered int64. recompute_schema runs unconditionally
again; only resolve_lambda_variables stays gated. The first-seat experiment
is exonerated (its revert stayed red) and stays reverted.
The facade factorial builder embeds the Factorial shadow so both doors
resolve one kernel (door-parity test green).
483/483 pins+HOF green on the release build; PERF numbers re-measured.

pins: sql-literal-typing-1/L-002, PERF-001, PERF-005

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* docs(sql-literal-typing-1): remediation round 1 gate lines

verify 0 (56 ok), clippy 0, repark-spark lib 1041/0, repark-python lib
75/0, facade 9108/0, parity 757/0, release native rebuilt. The 24 FNP-8
HOF pins broke mid-round on the recompute skip and are green after the
always-recompute fix.

pins: sql-literal-typing-1/L-001, L-002, L-003

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* test(sql-literal-typing-1): red pin for V-001 HOF parenthesized INT_MIN

transform/filter/transform_keys over -(2147483648) answer int widths
while the top-level door answers bigint (LIT2-SQL-03); both doors must
agree on bigint.

pins: sql-literal-typing-1/V-001

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* fix(sql-literal-typing-1): drop the Negative fold from the shared provisional helper

HigherOrderPreparation narrows lambda bodies and value constructors
through narrow_provisional_integer_literal, whose Negative arm still
folded -(2147483648) to INT while the top-level door answers bigint.
One narrowing path now: only the lexer-level negative token narrows.
77/77 pins green on the rebuilt release native.

pins: sql-literal-typing-1/V-001

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* docs(sql-literal-typing-1): round 3 V-001 gate lines

functions lib 738/0, spark lib 1041/0, clippy 0, release rebuilt,
-k slice 722/0. Orchestrator runs verify plus the full suites.

pins: sql-literal-typing-1/V-001

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

* docs(sql-literal-typing-1): orchestrator close-out — review verdicts, gate audit, rulings

Authored-By: Claude (claude-opus-5) <noreply@anthropic.com>

---------

+++ python/repark/tests/test_case_insensitive_conform.py
@@ -53,7 +53,7 @@ def test_merge_star_conforms_case_differing_source_by_name(spark: ReparkSession)
         {"id": 3, "name": "c"},
     ]
     # Value AND Arrow type on the export path (to_arrow), never show.
-    assert table.schema.field("id").type == pa.int64()
+    assert table.schema.field("id").type == pa.int32()
     assert table.schema.field("name").type == pa.string()
